Various refactorings.
`getcolumn()` is centralised, so that both `getcolumntd()` and `getcolumnth()` may call it.

`ordercolumn()` is modified to return a sorted array, rather than operating on the DOM tree directly. I think this was part of an optimisation for avoiding relatively slow DOM operations by caching the result.

Various short-lived variables are renamed for consistency, and a few things clarified.