A start on generating coverage reports for API implementations across various OSes.
The source for this information is formed of two parts: 
 * The API specifications, the symbols from which are extracted from the compiler's symbol table dump, when parsing the tspec-generated API interface headers. Note that these headers also contain OS-independent implementations of portable constructs, such as NULL's value (value, not representation).
 * OS-dependent implementations of varying percentages of those API specifications, taken from the API `.tl` token libraries, interrogated rather crudely courtesy of `tnc`. Eventually, I envisage that these will be delivered automatically by various build machines.

These sources are a first attempt at getting this information; they may not be entirely correct - that is, they may miss certian information, or mislead about certian information. Time will tell.

There is a lot of data to convey here; it's probable that several different aspects will be required. For now, I'm just showing API versus OS, subdivided into OS versions and API subsections. For sake of sanity, I've shown only summarised rows, and collapsed away the details, which are optionally expandable if you wish to see specific symbols.

This will probably take a fair amount of work before I'm happy with it. But then again, what doesn't?