A tentative implementation of a simple blog. This uses mod_kxslt to provide date indexes from the httpd query string, and presents appropriate entries through the timeline XSLT.
The timelined entries (not present in this commit) are consolidated at build time into one giant XML file.

This is work in progress, but the basic idea has taken shape. In particular the visual styling is just a rough sketch.

This page for `blog.libfsm.org/index.xhtml5` is a template, with XML processing instructions serving to mark where the dynamic content is to be substituted by `blog.xsl`. The bulk of gluing together the various generic XSLT and Javascript subsystems is done by `blog.xsl`, which is relatively small; it just brings in each system, and overrides the relevant templates with blog-specific things.

This commit is the culmination of quite a bit of work (both for this website and for others), and I am pleased with how this is turning out.