The website menu. This is presented as a finite state machine rendered in SVG, over which the user may hover and click to select their choice of content.
Transitions for viable routes to a given destiation are tagged on each node, and Javascript traverses these on-the-fly to provide an animation on page load which transitions through the FSM to arrive at the current page. The current page is identified by regexps matching against the URL, which are also tagged on SVG elements.

This idea was the starting point for the visual design of the entire site; it's been through quite a few iterations in order to get the aesthetic balance feeling right, which can be a bit difficult when communicating through graphviz.

The animation is work in progress; it's not quite right yet.

This took me bloody ages.