A HTTP REST interface for submitting comments. This is intended to be called by AJAX, and provides CORS (so that it may live under its own subdomain).
This validates and saves the comment; this is intended for use in `blog.libfsm.org`, but may also be applicable elsewhere (on changesets or tickets, for example). Validation is performed against a simple DTD permitting a subset of HTML markup.

As ever, this is work in progress; most of the various areas need some work, but the infrastructure is all there.