Implemented comment submission via AJAX. Comment previews have been reworked to share the same mechanism (this way the server-side validation is not needlessly reimplemented client-side). Markup in comments now has effect.
This commit touches a lot of areas, but the bulk of the work is in Javascript. Feedback on successful submission remains to be done. The XSLT changes are just supporting changes to provide for the DOM manipuation.

I am particularly pleased with how the `template.js` templates in `blog.libfsm.org/index.xhtml5` have turned out.