Bugfix for 'MM' having excess padding for Jeni's date:format-date() implementation.
The bug I see is for a format string of "M", zero-padded values are produced. I think "M" should format unpadded values, and "MM" should format zero-padded values (as indeed it does).

As I understand, the EXSLT date:format-date() format string is specified in terms of Java's SimpleDateFormat format. The documentation I found for that [2] is a little vague on the subject of "MM":

    Month: If the number of pattern letters is 3 or more, the month is interpreted as text; otherwise, it is interpreted as a number.

I believe the intention is for "MM" to produce zero-padded output (so an integer of 1 is formatted as the string "01"), and for a single "M" to produce unpadded output (so an integer of 1 is formatted as the string "1"). Her code for date.format-date.function.xsl seems to support that intention, as you test for $pattern-length = 2 explicitly:

{{{
                    <xsl:when test="$pattern-length = 2">
                       <xsl:value-of select="format-number($month, '00')" />
                    </xsl:when>
                    <xsl:otherwise>
                       <xsl:value-of select="$month" />
                    </xsl:otherwise>
}}

However in debugging I find that $month there is (sometimes?) bound to a zero-padded string of "08", rather than an integer value. So the effect of <value-of/> is to produce that padding verbatim.

Here's my patch:

iona% diff /tmp/date.format-date.function.xsl date.format-date.function.xsl
213c216
<                        <xsl:value-of select="$month" />
---
>                        <xsl:value-of select="format-number($month, '0')" />
iona*

This gives "8" for August, and "12" for December.

I don't know if this is the best solution, if indeed this is the expected behaviour. Maybe just casting to an integer would be less overhead, but I thought it made more sense to keep consistency with the format-number() in the clause above.