Abstraction for theme characteristics.
I've categorised characteristics of layouts (depending on the number of visible columns), such as whether a logo should go on the left vs. at the bottom, or if a sidebar is visible, and so on. Now `layout.styl` setting the various attributes for layouts (margins and whatnot) no longer encodes the criteria for column counts, which is nice if I change my mind about column widths.

I don't like the syntax i have for this interface, but at least the information all lives on the right side of the boundary. Maybe i can replace that with something nicer later on.